requirements.txt
flask==3.0.0
ldap3==2.9.1
python-dotenv==1.0.0
authlib==1.3.0
flask-login==0.6.3 
requests==2.28.1